Remove mean error for empty groups
mean_with_confints_groups fails if for a specific group value, the independent variable is entirely empty. The error is generated by the svyby function from the survey package, so this just checks to make sure there is at least some data in the independent variable, and if not, calculations for this group value are skipped entirely.